+++ frontend/app/_layout.tsx
@@ -2,9 +2,9 @@ import { Stack } from "expo-router";
 import "../styles/global.css";
 import React from "react";
 import { AuthProvider } from "@/contexts/AuthContext";
-
 export default function RootLayout() {
   return (
+    <AuthProvider>
     <Stack screenOptions={{
       headerShown: false,
     }}>
@@ -12,5 +12,6 @@ export default function RootLayout() {
       <Stack.Screen name="/auth/signUp" />
       <Stack.Screen name="/sign" />
     </Stack>
+    </AuthProvider>
   );
 }
